# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Flask JWT Authentication API
  version: 1.0.0
  description: Authentication and user management API using JWT access and refresh tokens.
servers:
  - url: http://127.0.0.1:5000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: API and database are healthy
  /api/v1/auth/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201": {description: User registered}
        "409": {description: Email already registered}
        "422": {description: Validation error}
  /api/v1/auth/login:
    post:
      summary: Authenticate and receive access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200": {description: Authentication succeeded}
        "401": {description: Invalid credentials}
        "423": {description: Account temporarily locked}
  /api/v1/auth/refresh:
    post:
      summary: Rotate the refresh token and create a new token pair
      security: [{bearerAuth: []}]
      responses:
        "200": {description: New access and refresh tokens created}
  /api/v1/auth/logout:
    post:
      summary: Revoke the supplied access or refresh token
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Token revoked}
  /api/v1/auth/logout-all:
    post:
      summary: Invalidate all tokens for the current user
      security: [{bearerAuth: []}]
      responses:
        "200": {description: All tokens invalidated}
  /api/v1/users/me:
    get:
      summary: Get the current user
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Current user}
    patch:
      summary: Update the current user name using a fresh access token
      security: [{bearerAuth: []}]
      responses:
        "200": {description: User updated}
  /api/v1/users/me/password:
    patch:
      summary: Change password and invalidate existing tokens
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Password changed}
  /api/v1/users:
    get:
      summary: List users (admin)
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Paginated users}
        "403": {description: Admin role required}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: {type: string, minLength: 2, maxLength: 120}
        email: {type: string, format: email}
        password: {type: string, minLength: 10, maxLength: 128}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}